refactor(core): move common types to `lib.rs`
